# modules/reservation/reservation-headless/reservation-headless-impl/rest-openapi.yaml
components:
  schemas:
...
    Amenity:
      description: Contains information about a room's available amenities (microphones, projector, etc.)
      properties:
        amenityId:
          description: The primary key for the Amenity entity.
          type: integer
          format: int64
          readOnly: true
        name:
          description: The amenity name.
          type: string
      required: 
        - name
      type: object
...
openapi: 3.0.1
info:
  description: "API for accessing Reservation details."
  license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "Headless Reservation"
  version: v1.0
paths:
...
  "/amenities":
    get:
      operationId: getAmenitiesPage
      tags: ["Amenity"]
      description: Retrieve the list of amenities. Results can be paginated, filtered, searched and sorted.
      parameters:
        - in: query
          name: filter
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Amenity"
                type: array
            application/xml:
              schema:
                items:
                  $ref: "#/components/schemas/Amenity"
                type: array
    post:
      operationId: postAmenity
      tags: ["Amenity"]
      description: Create a new amenity.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Amenity"
          application/xml:
            schema:
              $ref: "#/components/schemas/Amenity"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Amenity"
            application/xml:
              schema:
                $ref: "#/components/schemas/Amenity"
  "/amenities/{amenityId}":
    get:
      operationId: getAmenity
      tags: ["Amenity"]
      description: Retrieves the amenity via its ID.
      parameters:
        - name: amenityId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Amenity"
            application/xml:
              schema:
                $ref: "#/components/schemas/Amenity"
    put: 
      operationId: putAmenity
      tags: ["Amenity"]
      description: Updates and overwrites the amenity.
      parameters:
        - name: amenityId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Amenity"
          application/xml:
            schema:
              $ref: "#/components/schemas/Amenity"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Amenity"
            application/xml:
              schema:
                $ref: "#/components/schemas/Amenity"
    delete:
      operationId: deleteAmenity
      tags: ["Amenity"]
      description: Deletes the amenity and returns a 204 if the operation succeeds.
      parameters:
        - name: amenityId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: ""
          content:
            application/json: {}
            application/xml: {}
...